# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lead Management API",
    "version": "1.0.0",
    "description": "REST API for managing sales leads."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/settings": {
      "get": {
        "summary": "Get the current user's settings (created with defaults if absent)",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Settings object"
          }
        }
      },
      "put": {
        "summary": "Update the current user's settings",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "theme": {
                    "type": "string",
                    "enum": [
                      "light",
                      "dark"
                    ]
                  },
                  "pageSize": {
                    "type": "integer"
                  },
                  "showAvatars": {
                    "type": "boolean"
                  },
                  "confirmBeforeDelete": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated settings"
          }
        }
      }
    },
    "/api/leads/dashboard": {
      "get": {
        "summary": "Total and per-status lead counts (computed in the database)",
        "tags": [
          "Leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard stats"
          }
        }
      }
    },
    "/api/leads/reports": {
      "get": {
        "summary": "Pipeline report aggregates (all computed in the database)",
        "tags": [
          "Leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Report aggregates by status and source with derived KPIs"
          }
        }
      }
    },
    "/api/leads": {
      "get": {
        "summary": "List leads (server-side search, status filter, and pagination)",
        "tags": [
          "Leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "NEW",
                "CONTACTED",
                "QUALIFIED",
                "WON",
                "LOST"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated leads envelope"
          }
        }
      },
      "post": {
        "summary": "Create a lead",
        "tags": [
          "Leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Lead created"
          }
        }
      }
    },
    "/api/leads/{id}": {
      "get": {
        "summary": "Get a lead by id",
        "tags": [
          "Leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lead found"
          },
          "404": {
            "description": "Lead not found"
          }
        }
      },
      "put": {
        "summary": "Update a lead",
        "tags": [
          "Leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lead updated"
          },
          "404": {
            "description": "Lead not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a lead",
        "tags": [
          "Leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Lead deleted"
          },
          "404": {
            "description": "Lead not found"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Log in and receive a JWT",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT issued"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    }
  },
  "tags": []
}